brew-report-issue: report/close project issues.
This script has two modes: filing issues and closing them.

When filing issues the body is piped in over STDIN and the issue title
and repository specified on the command-line. To avoid noise: this
should probably only be used on dedicated repositories for these issues
as they may be autoposted frequently.

When closing issues all issues filed by the user will be closed on the
repository and a comment with the success state will be posted to each
of the issues.

GitHub API credentials are extracted based on access to HTTPS GitHub
repositories.

This tool is designed to behave similarly to how Boxen did and will be
used by projects for autoposting issues.