brew-report-issue: setup Git credentials.
`git credential fill` prompts for but does not save any credentials.
Instead of asking for them every time look for a credential helper, ask
for the credentials once, store them and invalidate them if
authentication failed. For edge-cases: tell people to use Strap which
sets this all up for them instead.