Set up nodenv/rbenv automatically
When calling the brew-bootstrap scripts, eval the nodenv/rbenv init steps
automatically instead of errorring out if the user hasn't done it themselves.

Erroring out causes script/bootstrap calls to fail; it's probably preferable to
warn the user to do this for their shell and keep going, rather than cause the
script which called the bootstrap step to fail.