bootstrap-rbenv-ruby: respect RBENV_VERSION
This resolves an inconsistency which would occur if
RBENV_VERSION was exported, and also makes it possible to use
that environment variable to specify the Ruby version to install.

If RBENV_VERSION is exported, the rbenv version-name check
will respect that environment variable; however, the RUBY_REQUESTED
check was hardcoded to use rbenv local, which only considers
.ruby-version. This improves that check by using RBENV_VERSION if
specified, and falls back to .ruby-version if not; that behaviour
matches the fallback preference used by rbenv version-name.