Respect HOMEBREW_PREFIX in brew-macos-vscode-codespaces
I have my Homebrew files in `/opt/homebrew` because I'm running on
an M1, but the script is hard-coded to assume `/usr/local`. As such,
the script tries to set the symlink from `/usr/local/etc/resolver` to
`/etc/resolver`, and it ends up being a non-existent source.

The fix is to use `$HOMEBREW_PREFIX` if set, and fallback to
`/usr/local` otherwise.

Further down the file, we also use the `$HOMEBREW_PREFIX` to prepend
`/opt/homebrew/bin` or `/usr/local/bin` to the `$PATH` rather than
hard-coding `/usr/local/bin`. In practice, however, this probably
doesn't matter because if whatever your `$HOMEBREW_PREFIX` is,
`$HOMEBREW_PREFIX/bin` is probably already in your `$PATH` by the time
you run `brew macos-vscode-codespaces`.