Fix extension installation in brew-macos-vscode-codespaces
While testing the previous commit, I noticed that this script isn't
actually installing any extensions. There are two problems.

First is that we can't assign a multiline string to the
`NEEDED_EXTENSIONS` variable like that. The `for` loop body was never
being entered, as a result. We instead have to use one of the techniques
listed here[0].

I would have used the `read` technique[1], as it avoids forking a
useless subshell, but it causes the script to die because we are using
`set -e` and `read` exits with a status of `1` at EOF, as explained in
this comment[2].

So, instead, using this alternative technique [3].

The second problem with the script is that we we feeding a non-existent
variable (`$EXTENSIONS`) into our `grep` invocation, meaning that we
failed to correctly detect an existing install. The fix is to use
`$INSTALLED_EXTENSIONS` instead.

Test plan: Copy the changed script to
`/opt/homebrew/Library/Taps/github/homebrew-bootstrap/cmd/` and run
`brew macos-vscode-codespaces`. I sprinkled some `echo` through the file
to see it enter the loop and exit appropriately. It only installs on the
first run.

[0]: https://stackoverflow.com/questions/1167746/how-to-assign-a-heredoc-value-to-a-variable-in-bash
[1]: https://stackoverflow.com/a/1655389
[2]: https://stackoverflow.com/questions/1167746/how-to-assign-a-heredoc-value-to-a-variable-in-bash#comment7650320_1655389
[2]: https://stackoverflow.com/a/1168084